+++ src/test/java/kitchenpos/application/TableGroupServiceTest.java
@@ -15,14 +15,13 @@ import kitchenpos.application.response.TableGroupResponse;
 import kitchenpos.dao.MenuCustomDao;
 import kitchenpos.dao.OrderCustomDao;
 import kitchenpos.dao.TableGroupCustomDao;
-import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.ordertable.NumberOfGuests;
 import kitchenpos.domain.product.Price;
-import kitchenpos.ui.request.OrderLineItemsRequest;
+import kitchenpos.application.request.OrderLineItemsRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
