+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -17,7 +17,6 @@ import kitchenpos.application.response.OrderTableResponse;
 import kitchenpos.dao.MenuCustomDao;
 import kitchenpos.dao.OrderCustomDao;
 import kitchenpos.dao.OrderTableDao;
-import kitchenpos.domain.Menu;
 import kitchenpos.domain.MenuGroup;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
@@ -25,7 +24,7 @@ import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.ordertable.NumberOfGuests;
 import kitchenpos.domain.product.Price;
-import kitchenpos.ui.request.OrderLineItemsRequest;
+import kitchenpos.application.request.OrderLineItemsRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
