+++ src/test/java/kitchenpos/ui/OrderRestControllerTest.java
@@ -13,8 +13,8 @@ import kitchenpos.application.response.OrderResponse;
 import kitchenpos.domain.Order;
 import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderStatus;
-import kitchenpos.ui.request.OrderLineItemsRequest;
-import kitchenpos.ui.request.OrderRequest;
+import kitchenpos.application.request.OrderLineItemsRequest;
+import kitchenpos.application.request.OrderRequest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
